General: *         Upgraded Steamworks SDK to v1.29 *         Fixed mod compatibility problem with Multiplayer Base that was introduced in September. *         In Hammer, while using the Vertex Tool, pressing CTRL+B will snap selected vertices to the grid.
Virtual Reality:
*         Mods that support virtual reality now need to have a line in gameinfo.txt that says “supportsvr 1”. This indicates to gameui and engine that certain UI should be enabled.
*         VR-enabled mods will now start up in VR mode when launched from Steam’s VR mode.

Windows:
*         Upgraded to Visual Studio 2013. If you need to build projects for VS 2010, add /2010 to your VPC command line.

OSX:
*         Upgraded to XCode 5.

+++ mp/src/tier1/generichash.cpp
@@ -364,7 +364,7 @@ uint32 MurmurHash2( const void * key, int len, uint32 seed )
 #define TOLOWERU( c ) ( ( uint32 ) ( ( ( c >= 'A' ) && ( c <= 'Z' ) )? c + 32 : c ) )
 uint32 MurmurHash2LowerCase( char const *pString, uint32 nSeed )
 {
-	int nLen = strlen( pString );
+	int nLen = ( int )strlen( pString );
 	char *p = ( char * ) stackalloc( nLen + 1 );
 	for( int i = 0; i < nLen ; i++ )
 	{
